project_localy/requirements.txt
nltk==3.9.1
numpy==1.26.4
gensim==4.3.3
pandas==2.2.3
scipy==1.10.1
scikit-learn==1.6.1
matplotlib==3.10.0
gitpython==3.1.44
seaborn==0.13.2
openpyxl>=3.0.0